# cursillo/swagger/dist/schema/swagger.json
{
  "swagger":"2.0",
  "info":{
    "description":"This is the Cursillo Swagger documentation.",
    "version":"1.0.0",
    "title":"Cursillo Swagger "
  },
  "host":"localhost",
  "basePath":":8080/api/",
...
  "schemes":[
    "http"
  ],
  "paths":{
...
    "/positions": {
      "get":{
        "tags":[
          "position"
        ],
        "summary":"Gets all positions.",
        "description":"",
        "operationId":"getPositions",
        "consumes":[
          "application/json"
        ],
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "in":"query",
            "name":"authorization",
            "description":"Access Token",
            "required":true,
            "type": "string"
          }
        ],
        "responses":{
          "400":{
            "description":"Invalid credentials."
          }
        },
        "security":[

        ]
      },
      "post":{
        "tags":[
          "position"
        ],
        "summary":"Creates a new position.",
        "description":"",
        "operationId":"createPosition",
        "consumes":[
          "application/json"
        ],
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "in":"query",
            "name":"authorization",
            "description":"Access Token",
            "required":true,
            "type": "string"
          },
          {
            "in":"body",
            "name":"body",
            "description":"Create a new position",
            "required":true,
            "schema":{
              "$ref":"#/definitions/Position"
            }
          }
        ],
        "responses":{
          "400":{
            "description":"Invalid credentials."
          }
        },
        "security":[

        ]
      }
    },
    "/positions/{positionId}": {
      "get":{
        "tags":[
          "position"
        ],
        "summary":"Return position with Id positionId.",
        "description":"",
        "operationId":"getPositionById",
        "consumes":[
          "application/json"
        ],
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "in":"query",
            "name":"authorization",
            "description":"Access Token",
            "required":true,
            "type": "string"
          },
          {
            "in":"path",
            "name":"positionId",
            "description":"Position Id",
            "required":true,
            "type": "integer",
            "form": "int32"
          }
        ],
        "responses":{
          "400":{
            "description":"Invalid credentials."
          }
        },
        "security":[

        ]
      },
      "put":{
        "tags":[
          "position"
        ],
        "summary":"Updates an existing position.",
        "description":"",
        "operationId":"updatePosition",
        "consumes":[
          "application/json"
        ],
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "in":"query",
            "name":"authorization",
            "description":"Access Token",
            "required":true,
            "type": "string"
          },
          {
            "in":"path",
            "name":"positionId",
            "description":"Id of position being updated.",
            "required":true,
            "schema":{
              "$ref":"#/definitions/Position"
            }
          },
          {
            "in":"body",
            "name":"body",
            "description":"Create a new position",
            "required":true,
            "schema":{
              "$ref":"#/definitions/Position"
            }
          }
        ],
        "responses":{
          "400":{
            "description":"Invalid credentials."
          }
        },
        "security":[

        ]
      },
      "delete":{
        "tags":[
          "position"
        ],
        "summary":"Delete position with Id positionId.",
        "description":"",
        "operationId":"deletePositionById",
        "consumes":[
          "application/json"
        ],
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "in":"query",
            "name":"authorization",
            "description":"Access Token",
            "required":true,
            "type": "string"
          },
          {
            "in":"path",
            "name":"positionId",
            "description":"Position Id",
            "required":true,
            "type": "integer",
            "form": "int32"
          }
        ],
        "responses":{
          "400":{
            "description":"Invalid credentials."
          }
        },
        "security":[

        ]
      }
    },
...
    "/weekends": {
      "get":{
        "tags":[
          "weekend"
        ],
        "summary":"gets all weekends.",
        "description":"",
        "operationId":"getweekends",
        "consumes":[
          "application/json"
        ],
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "in":"query",
            "name":"authorization",
            "description":"access token",
            "required":true,
            "type": "string"
          }
        ],
        "responses":{
          "400":{
            "description":"invalid credentials."
          }
        },
        "security":[

        ]
      },
      "post":{
        "tags":[
          "weekend"
        ],
        "summary":"Creates a new weekend.",
        "description":"",
        "operationId":"createWeekend",
        "consumes":[
          "application/json"
        ],
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "in":"query",
            "name":"authorization",
            "description":"access token",
            "required":true,
            "type": "string"
          },
          {
            "in":"body",
            "name":"body",
            "description":"Create a new weekend",
            "required":true,
            "schema":{
              "$ref":"#/definitions/Weekend"
            }
          }
        ],
        "responses":{
          "400":{
            "description":"invalid credentials."
          }
        },
        "security":[

        ]
      }
    },
    "/weekends/{weekendId}": {
      "get":{
        "tags":[
          "weekend"
        ],
        "summary":"return weekend with id weekendId.",
        "description":"",
        "operationId":"getWeekendbyId",
        "consumes":[
          "application/json"
        ],
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "in":"query",
            "name":"authorization",
            "description":"access token",
            "required":true,
            "type": "string"
          },
          {
            "in":"path",
            "name":"weekendId",
            "description":"weekend id",
            "required":true,
            "type": "integer",
            "form": "int32"
          }
        ],
        "responses":{
          "400":{
            "description":"invalid credentials."
          }
        },
        "security":[

        ]
      },
      "put":{
        "tags":[
          "weekend"
        ],
        "summary":"updates an existing weekend.",
        "description":"",
        "operationId":"updateWeekend",
        "consumes":[
          "application/json"
        ],
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "in":"query",
            "name":"authorization",
            "description":"access token",
            "required":true,
            "type": "string"
          },
          {
            "in":"path",
            "name":"weekendId",
            "description":"id of weekend being updated.",
            "required":true
          },
          {
            "in":"body",
            "name":"body",
            "description":"create a new weekend",
            "required":true,
            "schema":{
              "$ref":"#/definitions/Weekend"
            }
          }
        ],
        "responses":{
          "400":{
            "description":"invalid credentials."
          }
        },
        "security":[

        ]
      },
      "delete":{
        "tags":[
          "weekend"
        ],
        "summary":"delete weekend with id weekendid.",
        "description":"",
        "operationId":"deleteweekendbyid",
        "consumes":[
          "application/json"
        ],
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "in":"query",
            "name":"authorization",
            "description":"access token",
            "required":true,
            "type": "string"
          },
          {
            "in":"path",
            "name":"weekendid",
            "description":"weekend id",
            "required":true,
            "type": "integer",
            "form": "int32"
          }
        ],
        "responses":{
          "400":{
            "description":"invalid credentials."
          }
        },
        "security":[

        ]
      }
    },
...
    "/weekendPositions": {
      "get":{
        "tags":[
          "weekendPosition"
        ],
        "summary":"gets all weekendPositions.",
        "description":"",
        "operationId":"getweekendPositions",
        "consumes":[
          "application/json"
        ],
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "in":"query",
            "name":"authorization",
            "description":"access token",
            "required":true,
            "type": "string"
          }
        ],
        "responses":{
          "400":{
            "description":"invalid credentials."
          }
        },
        "security":[

        ]
      },
      "post":{
        "tags":[
          "weekendPosition"
        ],
        "summary":"Creates a new weekendPosition.",
        "description":"",
        "operationId":"createWeekendPosition",
        "consumes":[
          "application/json"
        ],
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "in":"query",
            "name":"authorization",
            "description":"access token",
            "required":true,
            "type": "string"
          },
          {
            "in":"body",
            "name":"body",
            "description":"Create a new weekendPosition",
            "required":true,
            "schema":{
              "$ref":"#/definitions/WeekendPosition"
            }
          }
        ],
        "responses":{
          "400":{
            "description":"invalid credentials."
          }
        },
        "security":[

        ]
      }
    },
    "/weekendPositions/{weekendPositionId}": {
      "get":{
        "tags":[
          "weekendPosition"
        ],
        "summary":"return weekendPosition with id weekendPositionId.",
        "description":"",
        "operationId":"getWeekendPositionbyId",
        "consumes":[
          "application/json"
        ],
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "in":"query",
            "name":"authorization",
            "description":"access token",
            "required":true,
            "type": "string"
          },
          {
            "in":"path",
            "name":"weekendPositionId",
            "description":"weekendPosition id",
            "required":true,
            "type": "integer",
            "form": "int32"
          }
        ],
        "responses":{
          "400":{
            "description":"invalid credentials."
          }
        },
        "security":[

        ]
      },
      "put":{
        "tags":[
          "weekendPosition"
        ],
        "summary":"updates an existing weekendPosition.",
        "description":"",
        "operationId":"updateWeekendPosition",
        "consumes":[
          "application/json"
        ],
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "in":"query",
            "name":"authorization",
            "description":"access token",
            "required":true,
            "type": "string"
          },
          {
            "in":"path",
            "name":"weekendPositionId",
            "description":"id of weekendPosition being updated.",
            "required":true
          },
          {
            "in":"body",
            "name":"body",
            "description":"create a new weekendPosition",
            "required":true,
            "schema":{
              "$ref":"#/definitions/WeekendPosition"
            }
          }
        ],
        "responses":{
          "400":{
            "description":"invalid credentials."
          }
        },
        "security":[

        ]
      },
      "delete":{
        "tags":[
          "weekendPosition"
        ],
        "summary":"delete weekendPosition with id weekendPositionid.",
        "description":"",
        "operationId":"deleteweekendPositionbyid",
        "consumes":[
          "application/json"
        ],
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "in":"query",
            "name":"authorization",
            "description":"access token",
            "required":true,
            "type": "string"
          },
          {
            "in":"path",
            "name":"weekendPositionid",
            "description":"weekendPosition id",
            "required":true,
            "type": "integer",
            "form": "int32"
          }
        ],
        "responses":{
          "400":{
            "description":"invalid credentials."
          }
        },
        "security":[

        ]
      }
    },
...
  },
  "securityDefinitions":{
    "petstore_auth":{
      "type":"oauth2",
      "authorizationUrl":"http://petstore.swagger.io/oauth/dialog",
      "flow":"implicit",
      "scopes":{
        "write:pets":"modify pets in your account",
        "read:pets":"read your pets"
      }
    },
    "api_key":{
      "type":"apiKey",
      "name":"api_key",
      "in":"header"
    }
  },
  "definitions":{
...
    "Position": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "required": "true"
        },
        "number": {
          "type": "number",
          "required": "true"
        },
        "description": {
          "type": "string"
        }
      }
    },
...
    "Weekend": {
      "type": "object",
      "properties": {
        "number": {
          "type": "number",
          "required": true
        },
        "description": {
          "type": "string"
        },
        "startDate": {
          "type": "string"
        },
        "endDate": {
          "type": "string"
        },
        "notes": {
          "type": "string"
        },
        "isCompleted": {
          "type": "boolean"
        },
        "locationId": {
          "type": "integer",
          "format": "int64",
          "required": true
        }
      }
    },
    "WeekendPosition": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "isActive": {
          "type": "boolean"
        },
        "teamId": {
          "type": "integer",
          "format": "int64",
          "required": true
        },
        "personId": {
          "type": "integer",
          "format": "int64",
          "required": true
        },
        "positionId": {
          "type": "integer",
          "format": "int64",
          "required": true
        }
      }
    },
...
  }
}